Add get_color_table service
The service is used by the camera_color_calibration.

Fix a bug with loading the color table from a file. When reading
binary unsigned numbers from a file ifstream.read() should be used,
otherwise the results are wrong.

+++ tools/camera_color_calibration/src/model.cpp
@@ -37,7 +37,7 @@ bool Model::SendTable() {
   PixelClass* table_ptr = reinterpret_cast<PixelClass*>(table_);
   for (size_t i = 0; i < kTableLen; ++i) {
     PixelClass c = table_ptr[i];
-    uint len = 1;
+    size_t len = 1;
     while (table_ptr[i + len] == c) {
       ++len;
       if (i + len == kTableLen) break;
